swipe without having to wating for  loading
