+++ src/core/Content/ContentDisplayer.jsx
@@ -7,10 +7,16 @@ import LinearProgress from "@material-ui/core/LinearProgress";
 import BottomBar from "./components/BottomBar";
 import Empty from "./components/Empty";
 import { makeStyles } from "@material-ui/core/styles";
+import { Swipeable } from "react-swipeable";
 
 const useStyles = makeStyles({
-  secondary:{
-    marginBottom:"40px"
+  secondary: {
+    marginBottom: "40px"
+  },
+  root: {
+    height: "100vh",
+    background: "#ffffff22",
+    borderTop: "1px solid #fff"
   }
 });
 
@@ -29,6 +35,18 @@ export default function ContentDisplayer({
   loadSubject
 }) {
   const classes = useStyles();
+
+  function swipe(isLeft) {
+    if (isLeft) {
+      if (PrimarySliderSelectedIndex + 1 < content.length) {
+        handlePrimeTabClick(PrimarySliderSelectedIndex + 1);
+      }
+    } else {
+      if (PrimarySliderSelectedIndex > 0) {
+        handlePrimeTabClick(PrimarySliderSelectedIndex - 1);
+      }
+    }
+  }
 
   return (
     <div>
@@ -64,23 +82,31 @@ export default function ContentDisplayer({
         {/******  display Secondary slider depending on the selected prime  ******/}
 
         <Grid item xs={12}>
-          {PrimarySliderSelectedIndex !== false &&
-            content[PrimarySliderSelectedIndex].value !== false && (
-              <SecondarySlide
-                subject={subject}
-                content={content[PrimarySliderSelectedIndex].value}
-                removeFromTodo={removeFromTodo}
-                addToTodo={addToTodo}
-                todo={todo}
-                isVideo={isVideo}
-                handleVideoPin={handleVideoPin}
-                className={classes.secondary}
-                handlePrimeTabClick={handlePrimeTabClick}
-                PrimarySliderSelectedIndex={PrimarySliderSelectedIndex}
-                primeSliderLength={content.length}
-              />
-            )}
+          <Swipeable
+            onSwipedRight={() => swipe(0)}
+            onSwipedLeft={() => swipe(1)}
+          >
+            <div className={classes.root}>
+              {PrimarySliderSelectedIndex !== false &&
+                content[PrimarySliderSelectedIndex].value !== false && (
+                  <SecondarySlide
+                    subject={subject}
+                    content={content[PrimarySliderSelectedIndex].value}
+                    removeFromTodo={removeFromTodo}
+                    addToTodo={addToTodo}
+                    todo={todo}
+                    isVideo={isVideo}
+                    handleVideoPin={handleVideoPin}
+                    className={classes.secondary}
+                    handlePrimeTabClick={handlePrimeTabClick}
+                    PrimarySliderSelectedIndex={PrimarySliderSelectedIndex}
+                    primeSliderLength={content.length}
+                  />
+                )}
+            </div>
+          </Swipeable>
         </Grid>
+
         <BottomBar
           divided={divided}
           loadSubject={loadSubject}
